at deser. error: add path of invalid field to msg
My modification is for deserialization in strict mode.
In error case, the path to the field with an illegal value is included in the error message. Therefore, the path (in JSON Path format) is stored in the variable cx_sy_move_cast_error->source_typename.
The exception cx_sy_move_cast_error was also used before, only the variable source_typename was not filled.